FossilOrigin-Name: e88c1ff4cc4ce1888d19fc48d99dc8ab5b9d69fdbcc0a0f2acf0632116b501c8

+++ src/http_socket.c
@@ -198,13 +198,13 @@ end_socket_open:
 ** Send content out over the open socket connection.
 */
 size_t socket_send(void *NotUsed, const void *pContent, size_t N){
-  size_t sent;
+  ssize_t sent;
   size_t total = 0;
   while( N>0 ){
     sent = send(iSocket, pContent, N, 0);
     if( sent<=0 ) break;
-    total += sent;
-    N -= sent;
+    total += (size_t)sent;
+    N -= (size_t)sent;
     pContent = (void*)&((char*)pContent)[sent];
   }
   return total;
